% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_mobile.R
\name{dauphin_mobile}
\alias{dauphin_mobile}
\alias{dauphin_landline}
\alias{dauphin_mobile_landline}
\alias{print.dauphin_mobile}
\title{Standardize Australian mobilie phone numbers}
\usage{
dauphin_mobile(mob, ignore_calling_code = NA)

dauphin_landline(
  landline,
  default_area_code = getOption("daiphin.default_landline", 1L)
)

dauphin_mobile_landline(
  mob,
  landline,
  default_area_code = getOption("dauphin.default_landline", 1L)
)

\method{print}{dauphin_mobile}(x, ...)
}
\arguments{
\item{mob, landline}{Character vectors in which mobile numbers and landline
numbers are expected.}

\item{ignore_calling_code}{\code{logical(1)} Whether to ignore the calling code
\code{+61} in the result.}

\item{x}{A vector, generally a character vector, in which mobile numbers
are expected.}

\item{...}{Arguments passed to other methods.}

\item{default_landline}{An integer between 1 and 10 giving, for
landline numbers with 8 digits, what area code should be set. By default,
it is \code{1L}, but users should set the option
\code{"dauphin.default_landline"} so to correctly standardize non-mobile
numbers.}
}
\value{
Mobile phone numbers are represented as integer vectors. International
calling prefixes extend the number beyond the representation of signed
integers. We use \code{raw} vectors for the international prefix, if required.


If \code{ignore_calling_code = TRUE}, the integer vector is returned.
Elements of \code{x} for which the mobile phone number could not be
extracted map to \code{NA_integer_} in the result.

If \code{ignore_calling_code = FALSE}, then a list is returned. The second element
of the list is the calling prefix.

If \code{ignore_calling_code = NA} then it is set to \code{TRUE} if \code{x}
appears to have international prefixes already.
}
\description{
Standardize Australian mobilie phone numbers
}
\examples{
dauphin_mobile("0400 123 456")
dauphin_mobile("+61400123456", ignore_calling_code = TRUE)

dauphin_mobile_landline("0424 123 456", "03 1234 5678")
dauphin_mobile_landline(c("0424 123 456", "03 1234 5678"),
                        c(NA, "0424 123 456"))

}
